+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `test_status` and `exception` fields to AppMap metadata.
 * Write AppMap file atomically, by writing to a temp file first and then moving it into place.
 * Remove printing of `Inventory.json` file.
+* Remove source code from `classMap`.
 
 # v0.43.0